rename products selectors files
